Make the eager argument list splitting heuristic more conservative. (#1700)
Some folks on the Flutter team pointed out that the previous rule is too aggressive and splits even simple common code like:

```dart
Text('Item 1', style: TextStyle(color: Colors.white));
```

That's probably simple enough to stay on one line. So this tweaks the heuristic to allow that to remain unsplit. Where the old heuristic split any argument list with a named argument that contained a nested named argument, this requires there to be at least *three* named arguments.

It's sort of an arbitrary cutoff, but it seems to do a good job when run on a large corpus.