Get the test updater script working again.
When I made the style changes language versions for Dart 3.8, I changed
how tests are written to allow a single test to have different expected
outputs at different language versions.

I was pressed for time, so that broke the automated test updater which
doesn't know how to handle versions. This gets it working again.

In the process, I decided to tweak how the final newline in test files
is handled. For years, I avoided having a trailing newline because I
though it would confuse the formatter. At some point that may have even
been true. But it turns out to be fine and it's more idiomatic to have a
trailing newline at the end of a text file.

The test updater now ensures that and I went ahead and normalized all
of the tests to have trailing newlines. There are no other changes to
the test files.