Re-order tests in increasing version order.
When I first added support for version-specific tests, I had an idea
that there would be a "default" way to format a given test and then one
or more degenerate "old" ways to format it. Concretely, the 3.7 style
would be the old way and 3.8 (and later) the new way.

I thought it would be easier to read if the default newest style was
closest to the input, so the tests were ordered like:

```
<<<
foo.bar();
>>> No version default latest style here.
foo.bar();
>>> 3.7 Older worse style here.
foo.bar();
```

This works OK when there are only two outputs, a single old degenerate
style and the default style. But there are an increasing number of tests
for new language features (null-aware elements, dot shorthands). Those
tests shouldn't be run at all on certain older versions. And at some
point, I'm sure we will end up making a further style tweak that causes
some test to have multiple older styles.

At that point, I think it's clearer if the test outputs are in
chronological/version order, like:

```
<<<
foo.bar();
>>> 3.8 First version at which this feature is supported.
foo.bar();
>>> 3.9 Later style tweak.
foo.bar();
>>> 3.10 Another style tweak.
foo.bar();
```

Handling multiple versions like this will require further changes to
the test runner. This commit only reorders the test outputs in version
order to minimize diff clutter. There are no meaningful changes here,
it's just the output of a hacked version of the test_updater which
normalizes the order.