Enable language version 3.10 and format dot shorthands.
This is using a workaround for https://github.com/dart-lang/sdk/issues/60840.
That issue is fixed and a new version of analyzer is published, but
unfortunately there is no corresponding version of test that works with
it so I can't upgrade yet.

The workaround is pretty harmless, so I'm fine with it. (The formatter
uses similar logic to handle commas which also aren't represented in the
AST nodes.)