Bump version to publish 3.1.1. (#1746)
The only change in this version is allowing 3.9 as a supported language
version. This was rolled into the SDK a couple of weeks ago, but not
published.

I figure I should publish this before rolling the supported version
again to 3.10 just in case any external needs 3.9 support.

I'll land this before landing #1745.