Remove special case handling of dot shorthands in null aware elements. (#1757)
We fixed the parser so that it can correctly handle:

```dart
[?.property]
```

Even if there is no space between `?` and `.`, it knows to treat it as
a dot shorthand inside a null-aware element here. So the formatter
doesn't need to force insert a space in there to avoid `?.` from being
tokenized as a null-aware access.