Remove uses of internal analyzer implementation.
In the very rare case that a user called formatStatement() and gave the
formatter something longer than a single statement, we need to throw a
FormatException but don't actually have a diagnostic from the analyzer
to use.

To deal with that, it used to import some internal implementation
libraries in order to conjure up a diagnostic. Instead, this has a local
implementation of the necessary public analyzer classes.

This will avoid dart_style being broken with analyzer 9.0.0 when that
internal code changes.